Extract OutputBuffer class
